move info into own file
